Improve clarity of authentication
When I first came to this section, I read "token" in the 3rd sentence as being the value of the Authorization header, and created a header value of "<username>:api_token". I've tried to re-word it slightly so that it is clearer that it should be "<api_token>:api_token".